refactor OptionsSelector for PageEditor
