+++ resource/js/components/PageEditor.js
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 import * as toastr from 'toastr';
 import { throttle, debounce } from 'throttle-debounce';
 
-import { EditorOptions } from './PageEditor/EditorOptionsSelector';
+import { EditorOptions } from './PageEditor/OptionsSelector';
 import Editor from './PageEditor/Editor';
 import Preview from './PageEditor/Preview';
